[feature/main-260717] 소개 페이지 접근 링크 변경
